Switch to iterative per-record program flow. (#17)
* Switch to iterative per-record program flow.

This change switches up the flow as an optimization for very large
workspaces (15,000 in my case). One issue is that when the script fails,
it can sometimes cost up to an hour in lost effort given the API
throttling restrictions.

In the old flow, the channel list would be processed in it's entirety in
steps. With the new flow, each channel is analyzed and archived in turn,
such that a failure does not result in significant loss of work.

Overview of changes:
1. The whitelist file is only read one time, rather than for each
   invocation of the method
2. Archive method main loop is removed, but loop body persists and is
   applied once per record
3. get_inactive_channels is renamed to is_channel_disused and operates
   on a single channel record
4. Admin report is send one time at the end of the script as a digest,
   instead of spamming the channel with thousands of messages (in my
   case).
5. When processing a channel, a single . is printed for progress
   tracking. The archived channel log messages are similar and still
   logged.

* Flake8

* Flake8 retry