Message templates (#21)
* Add very basic templating for the channel messages.
This allows users to override the message sent to a channel just before
it is archived. For example, my team sends a message with a link to the
support channel because only Admins can un/archive channels.

The only supported dynamic formatting is one %s for the configurable
number of days. Should anybody have more advanced use cases in the
future, this templating can be built out or replaced with a more full
featured external templating dependency.

* Add support for customizable archive messages